course/pyproject.toml
[project]
name = "course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.7",
    "openai>=2.14.0",
    "pydantic-ai>=1.39.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.0",
    "tqdm>=4.66.0",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
